refactor(macos): extract Navigation from AppModel into AppModel+Navigation.swift
